requirements.txt
openai
langchain
faiss-cpu
pymupdf
unstructured
python-docx
streamlit
python-dotenv
langchain-community
tiktoken